+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -1,3 +1,4 @@
 export 'instagram.dart';
+export 'linkedin.dart';
 export 'telegram.dart';
 export 'whatsapp.dart';