+++ test/deeplink_x_test.dart
@@ -71,6 +71,11 @@ void main() {
         expect(action, isA<AppAction>());
       });
 
+      test('LinkedIn', () {
+        final action = WhatsApp.open();
+        expect(action, isA<AppAction>());
+      });
+
       test('IOSAppStore', () {
         const action = IOSAppStore.open;
         expect(action, isA<AppAction>());